More Functionality added
Can now search from the search bar with a natural manner.
Cards are displayed properly and can be zoomed on if clicked.
Can Add cards to a temporary deck that cannot be yet saved.
The deck is now viewable on the bottom side of the screen.
General refactoring of code and removal of many magic numbers.